# ai_tnhn/ai-api-tnhn/docs/swagger.yaml
basePath: /api
definitions:
...
  models.ContractCategory:
    properties:
      code:
        description: Unique identifier
        example: CAT_DUYTRI
        type: string
      created_at:
        type: integer
      deleted_at:
        type: integer
      description:
        example: Các hợp đồng về duy trì hệ thống
        type: string
      drive_folder_id:
        example: 1abc2def3ghi
        type: string
      id:
        type: string
      level:
        description: 0 for root
        example: 0
        type: integer
      name:
        example: Hợp đồng Duy trì
        type: string
      order:
        example: 1
        type: integer
      parent_id:
        description: Empty string if it's a root category
        example: ""
        type: string
      path:
        description: Materialized Path, e.g., ",id1,id2,"
        example: ',id1,id2,'
        type: string
      status:
        example: true
        type: boolean
      updated_at:
        type: integer
    type: object
...
  web.ErrorResponse:
    properties:
      code:
        example: 400
        type: integer
      error:
        example: error message
        type: string
      status:
        example: error
        type: string
    type: object
  web.Response:
    properties:
      code:
        example: 200
        type: integer
      data: {}
      status:
        example: success
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Máy chủ API cho Hệ thống Giám sát Ngập lụt Hà Nội.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: API Hệ thống Giám sát Ngập lụt Hà Nội
  version: "1.0"
paths:
...
  /contract-category:
    get:
      description: Truy xuất danh sách các danh mục hợp đồng có phân trang
      parameters:
      - description: Số trang
        in: query
        name: page
        type: integer
      - description: Số bản ghi mỗi trang
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.Response'
            - properties:
                data:
                  properties:
                    data:
                      items:
                        $ref: '#/definitions/models.ContractCategory'
                      type: array
                    total:
                      type: integer
                  type: object
              type: object
      security:
      - BearerAuth: []
      summary: Danh sách danh mục hợp đồng
      tags:
      - Danh mục hợp đồng
    post:
      consumes:
      - application/json
      description: Tạo mới một danh mục để tổ chức và quản lý các hợp đồng
      parameters:
      - description: Dữ liệu danh mục
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/models.ContractCategory'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.ContractCategory'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Tạo mới một danh mục hợp đồng
      tags:
      - Danh mục hợp đồng
  /contract-category/{id}:
    delete:
      description: Loại bỏ một danh mục hợp đồng ra khỏi hệ thống theo ID
      parameters:
      - description: ID danh mục
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: boolean
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Xóa danh mục hợp đồng
      tags:
      - Danh mục hợp đồng
    get:
      description: Truy xuất thông tin chi tiết của một danh mục hợp đồng cụ thể
      parameters:
      - description: ID danh mục
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.ContractCategory'
              type: object
      security:
      - BearerAuth: []
      summary: Lấy thông tin danh mục theo ID
      tags:
      - Danh mục hợp đồng
    put:
      consumes:
      - application/json
      description: Cập nhật thông tin chi tiết của một danh mục hợp đồng hiện có
      parameters:
      - description: ID danh mục
        in: path
        name: id
        required: true
        type: string
      - description: Dữ liệu danh mục cập nhật
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/models.ContractCategory'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.ContractCategory'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Cập nhật danh mục hợp đồng
      tags:
      - Danh mục hợp đồng
  /contract-category/tree:
    get:
      description: Truy xuất toàn bộ danh mục dưới dạng cấu trúc cây phân cấp
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.Response'
            - properties:
                data:
                  items:
                    type: object
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: Lấy cấu trúc cây danh mục
      tags:
      - Danh mục hợp đồng
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"